improved usart config function

+++ BARE/Inc/lcd.h
@@ -14,7 +14,7 @@ void Write_String(char*);
 void Scroll_Right(void);
 void Scroll_Left(void);
 void Blink_Cursor(void);
-void Change_Baud_Rate(uint8_t);
+void Change_Baud_Rate(uint32_t);
 void Reset_Baud_Rate(void);
 void Backlight_OFF(void);
 void Backlight_ON(void);
